+++ graphistry/__init__.py
@@ -1,6 +1,6 @@
 from graphistry.pygraphistry import (  # noqa: E402, F401
     client_protocol_hostname, protocol, server,
-    register, login, refresh, api_token, verify_token,
+    register, privacy, login, refresh, api_token, verify_token,
     store_token_creds_in_memory,
     name, description,
     bind, style, addStyle, edges, nodes, graph, settings,